updated add to shelf and fix some other bugs

+++ lib/features/book_details/bloc/book_details_event.dart
@@ -53,7 +53,7 @@ class DownloadBook extends BookDetailsEvent {
   final String title;
   final String author;
   final String series;
-  final double seriesIndex;
+  final int seriesIndex;
   final String directory;
   final DownloadSchema schema;
